+++ client/src/components/atoms/Comment/Comment.style.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components"
+
+export const Commentlayout = styled.li`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 7px;
+  margin-top: 7px;
+`
