+++ client/src/components/atoms/Comment/index.ts
@@ -1,0 +1,3 @@
+import Comment from './Comment'
+
+export default Comment
